extract metadata form files to component
